end StockController 260721

+++ SmokeEnGrill.API/Models/StoreProduct.cs
@@ -5,8 +5,8 @@ namespace SmokeEnGrill.API.Models
         public int Id { get; set; }
         public int StoreId { get; set; }
         public int ProductId { get; set; }
-        public int Qty { get; set; }
-        public int PendingQty { get; set; }
+        public decimal Qty { get; set; }
+        public decimal PendingQty { get; set; }
         public Product Product { get; set; }
         public Store Store { get; set; }
     }
